# docs/openapi/openapi.bundle.yaml
openapi: 3.0.3
info:
  title: ORM Query Analyzer API
  version: 1.0.0
  description: |
    API for capturing, analyzing, and reporting slow database queries across ORM-based applications. Supports multiple platforms, projects, and integrations with Slack and AI-powered analysis.
servers:
  - url: /
    description: Default server
security:
  - AccessToken: []
...
paths:
...
  /api/v1/projects:
    get:
      tags:
        - Projects
      summary: List projects (paginated)
      operationId: projectList
      security:
        - AccessToken: []
      parameters:
        - $ref: '#/components/parameters/PaginatePage'
        - $ref: '#/components/parameters/PaginatePerPage'
        - $ref: '#/components/parameters/PaginateSort'
        - $ref: '#/components/parameters/PaginateOrder'
        - $ref: '#/components/parameters/PaginateSearch'
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum:
              - active
              - inactive
              - archived
          description: Filter by project status
      responses:
        '200':
          description: Paginated projects list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectPaginatedResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorizedResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbiddenResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorValidationResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServerResponse'
    post:
      tags:
        - Projects
      summary: Create a project
      operationId: projectCreate
      security:
        - AccessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - platform_id
              properties:
                name:
                  type: string
                  minLength: 2
                  maxLength: 100
                  example: My Project
                description:
                  type: string
                  maxLength: 500
                  nullable: true
                  example: A sample project
                status:
                  type: string
                  enum:
                    - active
                    - inactive
                    - archived
                  default: active
                platform_id:
                  type: string
                  format: uuid
                  example: 550e8400-e29b-41d4-a716-446655440000
                gitlab:
                  type: object
                  properties:
                    project_id:
                      type: integer
                      example: 12345
                    url:
                      type: string
                      format: uri
                      example: https://gitlab.com/org/project
                    group_id:
                      type: integer
                      example: 100
                    group_name:
                      type: string
                      example: my-org
                    default_branch:
                      type: string
                      example: main
                    visibility:
                      type: string
                      example: private
                slack_channel:
                  type: object
                  properties:
                    slack_channel_id:
                      type: string
                      example: C01234ABCDE
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/ProjectV1Response'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorizedResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbiddenResponse'
        '409':
          description: Conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorConflictResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorValidationResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServerResponse'
    delete:
      tags:
        - Projects
      summary: Delete projects by IDs
      operationId: projectDeleteByIds
      security:
        - AccessToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - ids
              properties:
                ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                  example:
                    - 550e8400-e29b-41d4-a716-446655440000
      responses:
        '200':
          description: Projects deleted
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        type: object
                        nullable: true
                        example: null
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorizedResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbiddenResponse'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorValidationResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServerResponse'
...
  /api/v1/gitlab/projects:
    get:
      tags:
        - Gitlab
      summary: List GitLab projects (paginated)
      operationId: gitlabProjectList
      security:
        - AccessToken: []
      parameters:
        - $ref: '#/components/parameters/PaginatePage'
        - $ref: '#/components/parameters/PaginatePerPage'
        - $ref: '#/components/parameters/PaginateSort'
        - $ref: '#/components/parameters/PaginateOrder'
        - $ref: '#/components/parameters/PaginateSearch'
      responses:
        '200':
          description: Paginated GitLab projects list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GitlabProjectPaginatedResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorizedResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbiddenResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServerResponse'
  /api/v1/gitlab/projects/{projectId}/files:
    get:
      tags:
        - Gitlab
      summary: Read a file from a GitLab project
      operationId: gitlabProjectGetFile
      security:
        - AccessToken: []
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
          description: GitLab project ID
          example: 706
        - name: file_path
          in: query
          required: true
          schema:
            type: string
          description: Path to the file in the repository
          example: src/main.ts
        - name: ref
          in: query
          required: false
          schema:
            type: string
            default: main
          description: Branch name, tag, or commit SHA
          example: main
      responses:
        '200':
          description: File content retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/BaseResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/GitlabFileV1Response'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorUnauthorizedResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorForbiddenResponse'
        '404':
          description: File or project not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorNotFoundResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorInternalServerResponse'
components:
  securitySchemes:
    AccessToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
    RefreshToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT refresh token
    ProjectApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Project API key for SDK integration
  schemas:
    BaseResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Success
        data:
          type: object
          nullable: true
...
    ErrorUnauthorizedResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Not Authorized
        error:
          type: string
          example: No authorization token provided
    ErrorValidationResponse:
      type: object
      required:
        - message
        - errors
      properties:
        message:
          type: string
          example: Validation Error
        errors:
          type: array
          items:
            type: object
            properties:
              path:
                type: string
                example: email
              messages:
                type: array
                items:
                  type: string
                example:
                  - Invalid email format
    ErrorInternalServerResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Internal Server Error
        error:
          type: string
          example: Internal Server Error
    PaginateMeta:
      type: object
      required:
        - page
        - per_page
        - total
        - total_page
      properties:
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 10
        total:
          type: integer
          example: 100
        total_page:
          type: integer
          example: 10
...
    ErrorForbiddenResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Forbidden Access
        error:
          type: string
          example: Insufficient permissions
    ErrorConflictResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Conflict
        error:
          type: string
          example: Resource already exists
    ErrorNotFoundResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string
          example: Not Found
        error:
          type: string
          example: Resource not found
...
    PlatformV1Response:
      type: object
      required:
        - id
        - framework
        - orm_provider
        - database_provider
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        framework:
          type: string
          example: NestJS
        orm_provider:
          type: string
          example: TypeORM
        database_provider:
          type: string
          example: PostgreSQL
...
    ProjectKeyV1Response:
      type: object
      required:
        - id
        - name
        - masked_key
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          type: string
          example: Production Key
        masked_key:
          type: string
          example: qa_****abcd
        plain_key:
          type: string
          nullable: true
          description: Only returned on creation
          example: qa_1234567890abcdef
        last_used_at:
          type: string
          format: date-time
          nullable: true
          example: '2026-01-01T10:00:00+07:00'
    ProjectGitlabV1Response:
      type: object
      required:
        - id
        - gitlab_project_id
        - gitlab_url
        - gitlab_group_id
        - gitlab_group_name
        - gitlab_default_branch
        - gitlab_visibility
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        gitlab_project_id:
          type: integer
          example: 12345
        gitlab_url:
          type: string
          example: https://gitlab.com/org/project
        gitlab_group_id:
          type: integer
          example: 100
        gitlab_group_name:
          type: string
          example: my-org
        gitlab_default_branch:
          type: string
          example: main
        gitlab_visibility:
          type: string
          example: private
    ProjectV1Response:
      type: object
      required:
        - id
        - name
        - description
        - status
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          type: string
          example: My Project
        description:
          type: string
          nullable: true
          example: A sample project
        status:
          type: string
          enum:
            - active
            - inactive
            - archived
          example: active
        platform:
          $ref: '#/components/schemas/PlatformV1Response'
        keys:
          type: array
          items:
            $ref: '#/components/schemas/ProjectKeyV1Response'
        project_gitlab:
          $ref: '#/components/schemas/ProjectGitlabV1Response'
    ProjectPaginatedResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                meta:
                  $ref: '#/components/schemas/PaginateMeta'
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/ProjectV1Response'
...
    GitlabProjectV1Response:
      type: object
      properties:
        id:
          type: integer
          example: 706
        name:
          type: string
          example: queryanalyzerapi
        name_with_namespace:
          type: string
          example: playground / orm-query-analyzer / queryanalyzerapi
        path:
          type: string
          example: queryanalyzerapi
        path_with_namespace:
          type: string
          example: playground/orm-query-analyzer/queryanalyzerapi
        description:
          type: string
          nullable: true
          example: ORM Query Analyzer API
        web_url:
          type: string
          example: https://gitlab.dot.co.id/playground/orm-query-analyzer/queryanalyzerapi
        default_branch:
          type: string
          example: main
        created_at:
          type: string
          format: date-time
          example: '2025-01-01T10:00:00+07:00'
        last_activity_at:
          type: string
          format: date-time
          example: '2026-04-17T10:00:00+07:00'
    GitlabProjectPaginatedResponse:
      allOf:
        - $ref: '#/components/schemas/BaseResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                meta:
                  $ref: '#/components/schemas/PaginateMeta'
                items:
                  type: array
                  items:
                    $ref: '#/components/schemas/GitlabProjectV1Response'
    GitlabFileV1Response:
      type: object
      properties:
        file_name:
          type: string
          example: package.json
        file_path:
          type: string
          example: package.json
        size:
          type: integer
          example: 2048
        encoding:
          type: string
          example: base64
        content_sha256:
          type: string
          example: abc123def456
        ref:
          type: string
          example: main
        content:
          type: string
          description: File content (base64 encoded or raw depending on GitLab API)
          example: eyJuYW1lIjoicXVlcnlhbmFseXplcmFwaSJ9
  parameters:
    PaginatePage:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Current page number
    PaginatePerPage:
      name: per_page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 10
      description: Number of items per page
    PaginateSort:
      name: sort
      in: query
      required: false
      schema:
        type: string
        default: updated_at
      description: Field name for sorting
    PaginateOrder:
      name: order
      in: query
      required: false
      schema:
        type: string
        enum:
          - ASC
          - DESC
        default: DESC
      description: Sort direction
    PaginateSearch:
      name: search
      in: query
      required: false
      schema:
        type: string
      description: Search keyword to filter results